refactor out packfieldmap and keyword
